Remove wrapping of step arguments in a List
This wrapping means that all arguments passed to a step are included in just one argument in the Bytecode that contains them in a List which doesn't work with the server.
IMPORTANT: Steps that have a params [] argument currently crash when the params[] is empty, e.g., when HasLabel is called with just one label. The reason is that the empty array will be sent to the server which doesn't expect this. So we have to add a check for such a params[] and only pass it to the Bytecode if it actually contains arguments.