Fix problem with params parameters
All arguments are now passed as one array to the Bytecode. This means
that steps with a params parameter wrap all of their arguments in just
one array, including the arguments from the params array.
In addition, the steps from the anoynmous traversal source check whether
their params parameter actually contains parameters before passing it to
the step in GraphTraversal to avoid passing an empty array that would
end up in the Bytecode.